fix: 확장자명 변경
