+++ src/App.jsx
@@ -1,36 +1,19 @@
-import { useEffect, useState } from "react";
-import { Header } from "./components/Header/header.jsx";
-import { Footer } from "./components/Footer/footer.jsx";
-import { Shared } from "./pages/shared.jsx";
-import { getFetch } from "./api/getFetch.js";
-import { getUserData } from "./api/getUserData.js";
-import { Folder } from "./pages/folder.jsx";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import "./app.css";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import "./App.css";
+import FolderPage from "./page/FolderPage";
+import SharedPage from "./page/SharedPage";
 
-export default function App() {
-  const [folder, setFolder] = useState(null);
-  const [user, setUser] = useState(null);
-
-  useEffect(() => {
-    const getData = async () => {
-      const data = await getFetch();
-      setFolder(data.folder);
-
-      const user = await getUserData();
-      setUser(user.data?.[0] ?? null);
-    };
-    getData();
-  }, []);
-
+function App() {
   return (
     <BrowserRouter>
-      {folder && <Header folder={folder} user={user} />}
       <Routes>
-        <Route path="/shared" element={<Shared />} />
-        <Route path="/folder" element={<Folder />} />
+        <Route path="/">
+          <Route path="/folder" element={<FolderPage />} />
+          <Route path="/shared" element={<SharedPage />} />
+        </Route>
       </Routes>
-      <Footer />
     </BrowserRouter>
   );
 }
+
+export default App;
