+++ src/components/Modal/CloseModalButton/index.jsx
