+++ src/components/SharedPageHeaderUserInfo/index.jsx
